Move reporting classes to their own directory and module
Just some tidyup that we wanted to do for a while
Helps a new user identify the entry point by removing clutter
